Максим, ты просил. Только подредач функцию боя, чтобы у тебя по кд не спавнился бой + учти, что тут без твоего кода
